revert moving out_of_slice_display, re-implement make_slice_request on vector layer
